% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgfun_dadosEntrada.R
\name{deckEntrada}
\alias{deckEntrada}
\title{Leitura dos decks de entrada}
\usage{
deckEntrada(
  deck,
  arqsAllowed = c("CARGAHIST", "EXOGENAHIST", "COMBINA", "FERIADOS", "PATAMARES",
    "EXOGENAPREV")
)
}
\arguments{
\item{deck}{Caminho dos arquivos ou lista contendo os data.frames das tabelas de entrada}

\item{arqsAllowed}{(opcional) lista de tabelas de entrada para filtro}
}
\value{
lista de data.frames com dados do deck de entrada do modelo
}
\description{
Leitura dos decks de entrada
}
