% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgfun_entradaDemandaMedia.R
\name{entradaDemandaMedia}
\alias{entradaDemandaMedia}
\title{Obtenha a entrada de dados para modelos de demanda media diaria}
\usage{
entradaDemandaMedia(
  serie,
  tipoEntrada = c("ML_DEFAULT", "ML_LOCAL", "TS_DEFAULT", "TS_FACTOR"),
  diaSemana = 1:7,
  diaAno = 1:366,
  lagsmodelo = c(1, 2, 7)
)
}
\arguments{
\item{serie}{serie de dados em demanda media a serem estruturados}

\item{tipoEntrada}{tipo de entrada do modelo: \cr
ML_DEFAULT - Entrada default para modelos de Machine Learning com Lags,Lags de Dias Especiais e Dummyes de Dias Especiais\cr
TS_DEFAULT - Entrada default para modelos de Series Temporais com Dummyes de: Dias Especiais, Dias da Semana e Meses\cr
TS_FACTOR - Entrada para modelos de Series Temporais com colunas de Dias Especiais, Dias da Semana e Meses em fatores\cr
ML_LOCAL - Entrada default para modelos de Machine Learning com Lags,Lags de Dias Especiais e Dummyes de Dias Especiais}

\item{diaSemana}{dias da semana para filtro}

\item{diaAno}{dias do ano para filtro}

\item{lagsmodelo}{lags de demanda e ou exogena}
}
\value{
data.table com os dados
}
\description{
Obtenha a entrada de dados para modelos de demanda media diaria
}
