% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgfun_entradaPerfil.R
\name{entradaPerfil}
\alias{entradaPerfil}
\title{Obtenha a entrada de dados para modelos horarios de perfil}
\usage{
entradaPerfil(
  serie,
  tipoEntrada = c("DEFAULT", "ML_PERFILFER"),
  hv = c(0, 1),
  diaSemana = 1:7,
  hora = 0:23,
  diaAno = 1:366,
  lagsmodelo = c(168, 336),
  lagsdsFer = NULL,
  tpFer = NULL,
  addHourFac = TRUE
)
}
\arguments{
\item{serie}{serie de dados a serem estruturados}

\item{tipoEntrada}{tipo de entrada do modelo:\cr
DEFAULT - Entrada default para modelo de perfil\cr
ML_PERFILFER - Entrada para perfil de feriados}

\item{hv}{indicador de horario de verao para filtro}

\item{diaSemana}{filtro de dias da semana de 1 a 7 (dom a sab)}

\item{hora}{filtro de horas do dia de 0 a 23}

\item{diaAno}{dias do ano para filtro}

\item{lagsmodelo}{lags dos modelos em horas}

\item{lagsdsFer}{lags em dia de semana em relacao a feriados}

\item{tpFer}{tipo de feriados}

\item{addHourFac}{adiciona coluna de dummy de horas}
}
\value{
data.table com os dados
}
\description{
Obtenha a entrada de dados para modelos horarios de perfil
}
