% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgfun_funcoesApoio.R
\name{obterModelos}
\alias{obterModelos}
\title{Obtem os modelos disponiveis para execucao}
\usage{
obterModelos()
}
\value{
Vetor com os modelos disponiveis para execucao
}
\description{
Obtem os modelos disponiveis para execucao
}
