% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prevcargadessem-package.R
\docType{package}
\name{prevcargadessem}
\alias{prevcargadessem}
\title{PrevCargaDESSEM}
\description{
Funcoes de execucao do modelo de previsao de carga PrevCargaDESSEM
}
\section{Funcoes para uso no pacote prevcargadessem}{

\itemize{
\item\code{\link{obterModelos}} - Obtem os modelos disponiveis para execucao
\item\code{\link{previsordessem}} - Execute o PrevCargaDESSEM
}
}

