% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgm_previsor.R
\name{previsordessem}
\alias{previsordessem}
\title{Execute o PrevCargaDESSEM}
\usage{
previsordessem(
  deck,
  dataPrev,
  horizonte = c("SEMOPER", "10DIAS"),
  modelos = obterModelos(),
  saida = c("CSV", "DESSEM", "PLOT"),
  dirSaida = ".",
  maxTreads = NULL
)
}
\arguments{
\item{deck}{Caminho para a pasta contendo os arquivos do deck de entrada (a pasta deve conter apenas os arquivos para a execucao) \cr
ou lista contento as tabelas nomeadas ("CARGAHIST","EXOGENAHIST","COMBINA","FERIADOS","PATAMARES","EXOGENAPREV")}

\item{dataPrev}{data do primeiro dia de previsao do modelo (D-1 do DESSEM)}

\item{horizonte}{"SEMOPER" (default) para previsoes ate o final da semana operativa de dataPrev\cr
"10DIAS" para previsoes ate 10 dias a frente, limitado pelo horizonte da variavel exogena prevista}

\item{modelos}{Modelos que se deseja rodar (default sao todos os disponiveis)\cr
A lista de modelos disponiveis pode ser obtida pela funcao \code{obterModelos()}\cr
Para execucao do modelo de combinacao, a opcao default deve ser mantida}

\item{saida}{"CSV" - escreve um arquivo csv no diretorio de saida\cr
"DESSEM" - escreve um arquivo txt com os campos dos blocos de carga do modelo DESSEM
"PLOT" - emite um plot com as previsoes}

\item{dirSaida}{Diretorio onde serao escritos os arquivos de saida do modelo}

\item{maxTreads}{Numero maximo de processos concorrentes do PrevCargaDESSEM (default - numero de nucleos da maquina menos 1)}
}
\value{
data.table com as previsoes feitas pelo do modelo
}
\description{
Execute o PrevCargaDESSEM
}
