% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkgfun_dadosEntrada.R
\name{TransformaDeck}
\alias{TransformaDeck}
\title{Transforma os decks de entrada em arquivos de Series Historicas e de Dados Diarios}
\usage{
TransformaDeck(
  deck,
  dataInicioPrev,
  horizonteprevisao = c("SEMOPER", "10DIAS")
)
}
\arguments{
\item{deck}{Deck de entrada}

\item{dataInicioPrev}{data de inicio da previsao}

\item{horizonteprevisao}{Horizonte de previsao Semana Operativa ou 10 dias c("SEMOPER","10DIAS")}
}
\value{
Dados para entrada de modelos
}
\description{
Transforma os decks de entrada em arquivos de Series Historicas e de Dados Diarios
}
